requirements.txt
numpy>=1.21.0
pandas>=1.3.0
openai>=1.0.0
python-dotenv>=0.19.0
langgraph>=0.0.10
scikit-learn>=0.24.2
statsmodels>=0.13.0 